Add a -N option to pkg_delete to delete registration information for a package but not its files. Override PKG_DELETE when building this package so -N can be used to upgrade an existing install. Bump version to 20041101.

+++ pkgtools/pkg_install/files/lib/plist.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: plist.c,v 1.10 2004/08/06 16:57:03 jlam Exp $	*/
+/*	$NetBSD: plist.c,v 1.11 2004/11/02 00:10:15 erh Exp $	*/
 
 #if HAVE_CONFIG_H
 #include "config.h"
@@ -11,7 +11,7 @@
 #if 0
 static const char *rcsid = "from FreeBSD Id: plist.c,v 1.24 1997/10/08 07:48:15 charnier Exp";
 #else
-__RCSID("$NetBSD: plist.c,v 1.10 2004/08/06 16:57:03 jlam Exp $");
+__RCSID("$NetBSD: plist.c,v 1.11 2004/11/02 00:10:15 erh Exp $");
 #endif
 #endif
 
@@ -335,7 +335,7 @@ write_plist(package_t *pkg, FILE * fp, char *realprefix)
  * run it too in cases of failure.
  */
 int
-delete_package(Boolean ign_err, Boolean nukedirs, package_t *pkg)
+delete_package(Boolean ign_err, Boolean nukedirs, package_t *pkg, Boolean NoDeleteFiles)
 {
 	plist_t *p;
 	char   *Where = ".", *last_file = "";
@@ -365,6 +365,8 @@ delete_package(Boolean ign_err, Boolean nukedirs, package_t *pkg)
 			break;
 
 		case PLIST_UNEXEC:
+			if (NoDeleteFiles)
+				break;
 			format_cmd(tmp, sizeof(tmp), p->name, Where, last_file);
 			if (Verbose)
 				printf("Execute `%s'\n", tmp);
@@ -381,6 +383,8 @@ delete_package(Boolean ign_err, Boolean nukedirs, package_t *pkg)
 				warnx("attempting to delete directory `%s' as a file\n"
 				    "this packing list is incorrect - ignoring delete request", tmp);
 			} else {
+				int     restored = 0;	/* restored from preserve? */
+
 				if (p->next && p->next->type == PLIST_COMMENT) {
 					if (strncmp(p->next->name, CHECKSUM_HEADER, ChecksumHeaderLen) == 0) {
 						char   *cp, buf[LegibleChecksumLen];
@@ -430,11 +434,9 @@ delete_package(Boolean ign_err, Boolean nukedirs, package_t *pkg)
 						}
 					}
 				}
-				if (Verbose)
+				if (Verbose && !NoDeleteFiles)
 					printf("Delete file %s\n", tmp);
-				if (!Fake) {
-					int     restored = 0;	/* restored from preserve? */
-
+				if (!Fake && !NoDeleteFiles) {
 					if (delete_hierarchy(tmp, ign_err, nukedirs))
 						fail = FAIL;
 					if (preserve && name) {
@@ -450,7 +452,9 @@ delete_package(Boolean ign_err, Boolean nukedirs, package_t *pkg)
 							}
 						}
 					}
+				}
 
+				if (!Fake) {
 					if (!restored) {
 #ifdef PKGDB_DEBUG
 						printf("pkgdb_remove(\"%s\")\n", tmp);	/* HF */
@@ -471,6 +475,9 @@ delete_package(Boolean ign_err, Boolean nukedirs, package_t *pkg)
 			break;
 
 		case PLIST_DIR_RM:
+			if (NoDeleteFiles)
+				break;
+
 			(void) snprintf(tmp, sizeof(tmp), "%s/%s", Where, p->name);
 			if (fexists(tmp)) {
 			    if (!isdir(tmp)) {
